Refactor common vue config, code split components
