+++ assets/js/helpers/storage.js
@@ -5,7 +5,7 @@
  * @link  https://developer.mozilla.org/en-US/docs/Web/API/Web_Storage_API/Using_the_Web_Storage_API
  * @param {string} type
  */
-function storageAvailable(type) {
+export function storageAvailable(type) {
     let storage;
     try {
         storage = window[type];
@@ -24,7 +24,3 @@ function storageAvailable(type) {
         );
     }
 }
-
-module.exports = {
-    storageAvailable
-};
