+++ assets/js/vue-components/vue-config.js
@@ -1,0 +1,3 @@
+export const VueConfig = {
+    delimiters: ['<%', '%>']
+};
